REFACTORING & SEARCH
/search/ is set up.. once you go into omens and update the  public static function findOmensBySearch(string $string){} , it should work

(and work with AJAX)

I've refactored OmenCollection and all of the Taxonomy pages

I've refactored some display pages to simplify code, and to use variables passed from routing

I've tried to move logic out of template files, and into index.php

I've updated the routing file

I've laid the groundwork for the AJAX to return JSON rather than HTML. I haven't implemented it yet, because it requires more javascript than I want to look at right now.

I've updated the omen listing page, so that when you click on the taxonomy, it shows in the URL bar (I don't know if this is good or not)

IMPORTANT::
I broke the ability to select or deselect an omen, when you're logged in. I don't know how I broke it, so I'd appreciate you taking a look at it.